requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
python-dotenv==1.0.1
joblib==1.4.2
sqlalchemy==2.0.36
